requirements.txt
Flask==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
psycopg2-binary==2.9.5
Werkzeug==2.2.2
WTForms-SQLAlchemy==0.3
flask-paginate==2024.4.12
gunicorn==20.1.0
python-dotenv==0.21.0
email-validator==1.3.1
